Updating de Fibonacci Equation
In this update a new function was added "equationR_fibonacci" that implement the recurrence relation of Fibonacci: Adding in the return of this function the method "Math.round()" to display an integer number.
This modification was added because in a test, ingresing the number 24, the number of the Fibonacci sequence in this position is "46368", and the method display "46367.99999999999".

In the Folder "Codigo" a new folder "Observaciones" was created, in this there was a screenshot with the 24's dilema.

Also, in the validation method, was added the validation of an integer value.